1.0/1/G/details_test.go
package main

import (
	"github.com/stretchr/testify/assert"
	"io"
	"strings"
	"testing"
)

func TestProblem(t *testing.T) {

	tests := []struct {
		name  string
		input string
		want  string
	}{
		{
			name:  "example test #1",
			input: "10 5 2",
			want:  "4",
		},
		{
			name:  "example test #2",
			input: "13 5 3",
			want:  "3",
		},
		{
			name:  "example test #3",
			input: "14 5 3",
			want:  "4",
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			var reader io.Reader = strings.NewReader(tt.input)
			result := problem(reader)
			assert.Equal(t, tt.want, result)
		})
	}

}
